+++ R/resample_est.R
@@ -6,7 +6,7 @@
 #' @param X object of class 'extended_selection_table' (see \code{\link{selection_table}}).
 #' @param samp.rate Numeric vector of length 1 with the sampling rate (in kHz) for output files. Default is \code{NULL}.
 #' @param bit.depth Numeric vector of length 1 with the dynamic interval (i.e. bit depth) for output files.
-#' @param sox Logical to control whether \href{http://sox.sourceforge.net/sox.html}{sox} is used internally for resampling. Sox must be installed. Default is \code{FALSE}. sox is a better option if having aliasing issues after resampling.
+#' @param sox Logical to control whether \href{http://sox.sourceforge.net/sox.html}{SOX} is used internally for resampling. Sox must be installed. Default is \code{FALSE}. \href{http://sox.sourceforge.net/sox.html}{SOX} is a better option if having aliasing issues after resampling.
 #' @param avoid.clip Logical to control whether the volume is automatically 
 #' adjusted to avoid clipping high amplitude samples when resampling. Ignored if
 #'  '\code{sox = FALSE}. Default is \code{TRUE}.
@@ -18,8 +18,8 @@
 #' @name resample_est
 #' @details This function aims to simplify the process of homogenizing sound 
 #' files (sampling rate and bit depth). This is a necessary step before running 
-#' any further (bio)acoustic analysis.  
-#'   
+#' any further (bio)acoustic analysis. Either \href{http://sox.sourceforge.net/sox.html}{SOX} (if \code{sox = TRUE}) or
+#'  or the \href{https://cran.r-project.org/package=bioacoustics}{bioacoustics package} (if \code{sox = FALSE}) should be installed.
 #' @examples
 #' \dontrun{
 #' # Set temporary working directory
@@ -49,8 +49,8 @@ resample_est <- function(X, samp.rate = 44.1, bit.depth = 16, sox = FALSE, avoid
 {
   
   # error message if bioacoustics is not installed
-  if (!requireNamespace("bioacoustics",quietly = TRUE))
-    stop("must install 'bioacoustics' to use mp32wav()")
+  if (!requireNamespace("bioacoustics",quietly = TRUE) & !sox)
+    stop("must install 'bioacoustics' to use mp32wav() when 'sox = FALSE'")
   
   # reset working directory 
   wd <- getwd()